feat:#31 Board Authentication

+++ src/main/java/com/nonangbie/board/service/BoardService.java
@@ -6,6 +6,8 @@ import com.nonangbie.boardLike.repository.BoardLikeRepository;
 import com.nonangbie.exception.BusinessLogicException;
 import com.nonangbie.exception.ExceptionCode;
 import com.nonangbie.member.entity.Member;
+import com.nonangbie.member.repository.MemberRepository;
+import com.nonangbie.utils.ExtractMemberEmail;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -21,38 +23,42 @@ import java.util.Optional;
 @Service
 @Transactional
 @RequiredArgsConstructor
-public class BoardService {
+public class BoardService extends ExtractMemberEmail {
     private final BoardRepository repository;
+    private final MemberRepository memberRepository;
     private final BoardLikeRepository boardLikeRepository;
     public Board createBoard(Board board, Authentication authentication) {
-        //시큐리티 적용 시 내용 추가 예정
-        Member member = new Member();
+        Member member = extractMemberFromAuthentication(authentication, memberRepository);
         board.setMember(member);
         return repository.save(board);
     }
     public Board updateBoard(Board board,long boardId,Authentication authentication) {
-        //시큐리티 적용 시 내용 추가 예정
-        Member member = new Member();
+        Member member = extractMemberFromAuthentication(authentication, memberRepository);
         Board findBoard = repository.findById(boardId)
                 .orElseThrow(() -> new BusinessLogicException(ExceptionCode.BOARD_NOT_FOUND));
         if(!Objects.equals(findBoard.getMember(), member))
             throw new BusinessLogicException(ExceptionCode.ACCESS_DENIED);
+        board.setMember(member);
         board.setLikeCount(findBoard.getLikeCount());
         board.setBoardId(findBoard.getBoardId());
         return repository.save(board);
     }
 
     public void deleteBoard(long boardId,Authentication authentication) {
-//        extractMemberFromAuthentication(authentication);
         Board findBoard = findVerifiedBoard(boardId, authentication);
+        String email = (String) authentication.getPrincipal();
+        if(!findBoard.getMember().getEmail().equals(email)) {
+            System.out.println("@@@@@@@Not Match@@@@@@");
+            System.out.println("@@@@@@@Not Match@@@@@@");
+            throw new BusinessLogicException(ExceptionCode.ACCESS_DENIED);
+        }
+        System.out.println("@@@@@@@###Match###@@@@@@");
         repository.delete(findBoard);
-//        if(!findBoard.getMember().getEmail().equals(authentication.getName())) {
-//            throw new BusinessLogicException(ExceptionCode.UNAUTHORIZED_MEMBER);
-//        }
 //        findBoard.setBoardStatus(Board.BoardStatus.BOARD_DELETED);
 //        boardRepository.save(findBoard);
     }
     public Page<Board> findBoardsByType(String type, String sort, int page, int size,Authentication authentication) {
+        extractMemberFromAuthentication(authentication, memberRepository);
         Sort sortBy;
         Board.MenuCategory menuCategory;
         try {
@@ -63,9 +69,9 @@ public class BoardService {
                 case "LIKE_DESC":
                     sortBy = Sort.by("likeCount").descending(); break;
                 case "CREATED_AT_ASC":
-                    sortBy = Sort.by("CREATED_AT").ascending(); break;
+                    sortBy = Sort.by("boardId").ascending(); break;
                 case "CREATED_AT_DESC":
-                    sortBy = Sort.by("CREATED_AT").descending(); break;
+                    sortBy = Sort.by("boardId").descending(); break;
                 default:
                     throw new IllegalArgumentException("Invalid sort type: " + sort);
             }
@@ -77,6 +83,7 @@ public class BoardService {
     }
 
     public boolean likeBoard(long boardId,Authentication authentication) {
+        extractMemberFromAuthentication(authentication, memberRepository);
         Board findBoard = repository.findById(boardId)
                 .orElseThrow(() -> new BusinessLogicException(ExceptionCode.BOARD_NOT_FOUND));
         findBoard.setLikeCount(findBoard.getLikeCount() + 1);
@@ -85,6 +92,7 @@ public class BoardService {
 
     @Transactional(readOnly = true)
     public Board findVerifiedBoard(long boardId, Authentication authentication){
+        extractMemberFromAuthentication(authentication, memberRepository);
         Optional<Board> findBoard =
                 repository.findById(boardId);
         Board result = findBoard.orElseThrow(()->
